Initial draft for v2
This commit is an initial draft for v2 of the library.

The library now has an error for each of the specified HTTP status codes
defined in the various RFCs.

We also made REST/JSON a higher priority, e.g. by removing
`BadRequestError` which returned a HTML page and now is the error type
for `400 Bad Request`.

We also removed the gRPC mapping, because there is no simple bijective
function to map gRPC error codes to HTTP status codes and vice versa.